+++ demo-snippets/vue/Ripples.vue
@@ -1,20 +1,18 @@
 <template>
-  <Page>
-    <ActionBar :title="title">
-      <NavigationButton text="Back"
-                        android.systemIcon="ic_menu_back"
-                        @tap="onNavigationButtonTap"></NavigationButton>
-    </ActionBar>
-    <StackLayout>
-      <MDRipple backgroundColor="gray" rippleColor="green" width="100" height="100" borderRadius="10"/>
-      <MDRipple width="100" height="100" borderRadius="10"/>
-    </StackLayout>
-  </Page>
+    <Page>
+        <ActionBar :title="title">
+            <NavigationButton text="Back" android.systemIcon="ic_menu_back" @tap="onNavigationButtonTap"></NavigationButton>
+        </ActionBar>
+        <StackLayout>
+            <MDRipple backgroundColor="gray" rippleColor="green" width="100" height="100" borderRadius="10" />
+            <MDRipple width="100" height="100" borderRadius="10" />
+        </StackLayout>
+    </Page>
 </template>
 
 <script lang="ts">
 import * as frameModule from '@nativescript/core/ui/frame';
-import Vue from 'vue'
+import Vue from 'vue';
 
 export const title = 'Ripples sample';
 
@@ -23,7 +21,7 @@ export default Vue.extend({
 
     data() {
         return {
-            title: title
+            title
         };
     },
 
@@ -31,6 +29,6 @@ export default Vue.extend({
         onNavigationButtonTap() {
             frameModule.Frame.topmost().goBack();
         }
-    },
-})
+    }
+});
 </script>