+++ packages/core-tabs/platforms/android/java/com/nativescript/material/core/TabsBar.java
@@ -429,14 +429,17 @@ public class TabsBar extends HorizontalScrollView {
                 return;
             }
             mTabStrip.onTabsViewPagerPageChanged(position, positionOffset);
+            if (positionOffset == 0 && mScrollState == androidx.viewpager.widget.ViewPager.SCROLL_STATE_SETTLING) {
+                scrollToTab(position, 0);
+            }
         }
 
         @Override
-        public void onPageScrollStateChanged(int state) {
+        public void onPageScrollStateChanged (int state) {
             super.onPageScrollStateChanged(state);
             mIgnoreScrollEventsUntilNextStateChange = false;
+            mScrollState = state;
         }
-
     }
 
     private class TabClickListener implements OnClickListener {