+++ src/textfield/textfield.ios.old.ts
@@ -32,7 +32,6 @@ import {
     placeholderColorProperty
 } from '@nativescript/core';
 import { textProperty } from '@nativescript/core/ui/text-base';
-import { layout } from '@nativescript/core/utils';
 import { TextFieldBase } from './textfield.common';
 
 @NativeClass
@@ -181,8 +180,8 @@ export class TextField extends TextFieldBase {
     }
     private mCanAutoSize = false;
     public onMeasure(widthMeasureSpec: number, heightMeasureSpec: number): void {
-        const widthMode = layout.getMeasureSpecMode(widthMeasureSpec);
-        this.mCanAutoSize = widthMode !== layout.EXACTLY;
+        const widthMode = Utils.layout.getMeasureSpecMode(widthMeasureSpec);
+        this.mCanAutoSize = widthMode !== Utils.layout.EXACTLY;
         super.onMeasure(widthMeasureSpec, heightMeasureSpec);
     }